report_generator/requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Email
aiosmtplib>=3.0.0

# Environment
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
